extract derivation types into separate module
